refactor arena launcher
